requirements.txt
# Core dependencies
# Add your project dependencies here
azure-ai-projects>=1.0.0b6
azure-identity>=1.12.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
